commentsScreen > Clean up the code
