+++ app/build.gradle
@@ -33,7 +33,8 @@ dependencies {
 
     implementation 'com.squareup.retrofit2:retrofit:2.2.0'
     implementation 'com.squareup.retrofit2:converter-simplexml:2.2.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.2.0'
+    implementation ('com.squareup.retrofit2:converter-gson:2.2.0')
+
 
     implementation 'com.nostra13.universalimageloader:universal-image-loader:1.9.5'
