+++ app/src/main/java/com/example/vmedvediev/redditapp/ImageLoaderManager.kt
@@ -1,0 +1,60 @@
+package com.example.vmedvediev.redditapp
+
+import android.content.Context
+import android.graphics.Bitmap
+import android.view.View
+import android.widget.ImageView
+import android.widget.ProgressBar
+import com.nostra13.universalimageloader.cache.memory.impl.WeakMemoryCache
+import com.nostra13.universalimageloader.core.DisplayImageOptions
+import com.nostra13.universalimageloader.core.ImageLoader
+import com.nostra13.universalimageloader.core.ImageLoaderConfiguration
+import com.nostra13.universalimageloader.core.assist.FailReason
+import com.nostra13.universalimageloader.core.assist.ImageScaleType
+import com.nostra13.universalimageloader.core.display.FadeInBitmapDisplayer
+import com.nostra13.universalimageloader.core.listener.ImageLoadingListener
+
+object ImageLoaderManager {
+
+    fun setupImageLoader(context: Context) {
+        val defaultOptions = DisplayImageOptions.Builder()
+                .cacheOnDisk(true).cacheInMemory(true)
+                .imageScaleType(ImageScaleType.EXACTLY)
+                .displayer(FadeInBitmapDisplayer(300)).build()
+
+        val config = ImageLoaderConfiguration.Builder(context)
+                .defaultDisplayImageOptions(defaultOptions)
+                .memoryCache(WeakMemoryCache())
+                .discCacheSize(100 * 1024 * 1024).build()
+
+        ImageLoader.getInstance().init(config)
+    }
+
+    fun showImage(context: Context, imageUrl: String, imageView: ImageView, progressBar: ProgressBar) {
+        val imageLoader = ImageLoader.getInstance()
+        val defaultImage = context.resources.getIdentifier("@drawable/reddit_alien", null, context.packageName)
+        val options = DisplayImageOptions.Builder().cacheInMemory(true)
+                .cacheOnDisk(true).resetViewBeforeLoading(true)
+                .showImageForEmptyUri(defaultImage)
+                .showImageOnFail(defaultImage)
+                .showImageOnLoading(defaultImage).build()
+
+        imageLoader.displayImage(imageUrl, imageView, options, object : ImageLoadingListener {
+            override fun onLoadingComplete(imageUri: String?, view: View?, loadedImage: Bitmap?) {
+                progressBar?.visibility = View.GONE
+            }
+
+            override fun onLoadingStarted(imageUri: String?, view: View?) {
+                progressBar?.visibility = View.VISIBLE
+            }
+
+            override fun onLoadingCancelled(imageUri: String?, view: View?) {
+                progressBar?.visibility = View.GONE
+            }
+
+            override fun onLoadingFailed(imageUri: String?, view: View?, failReason: FailReason?) {
+                progressBar?.visibility = View.GONE
+            }
+        })
+    }
+}
