+++ app/src/main/java/com/example/vmedvediev/redditapp/comments/CommentsActivity.kt
@@ -15,6 +15,11 @@ import android.widget.ProgressBar
 import android.widget.Toast
 import com.example.vmedvediev.redditapp.Account.LoginActivity
 import com.example.vmedvediev.redditapp.FeedAPI
+import com.example.vmedvediev.redditapp.ImageLoaderManager.setupImageLoader
+import com.example.vmedvediev.redditapp.ImageLoaderManager.showImage
+import com.example.vmedvediev.redditapp.NetworkManager.BASE_URL
+import com.example.vmedvediev.redditapp.NetworkManager.initGsonRetrofit
+import com.example.vmedvediev.redditapp.NetworkManager.initXmlRetrofit
 import com.example.vmedvediev.redditapp.R
 import com.example.vmedvediev.redditapp.WebViewActivity
 import com.example.vmedvediev.redditapp.XmlExtractor
@@ -45,16 +50,10 @@ class CommentsActivity : AppCompatActivity() {
 
     companion object {
         private const val TAG = "CommentsActivity"
-        private const val BASE_URL = "https://www.reddit.com/r/"
-        private const val COMMENT_URL = "https://www.reddit.com/api/"
     }
 
     private lateinit var postId: String
     private lateinit var postUrl: String
-    private lateinit var postThumbnailUrl: String
-    private lateinit var postTitle: String
-    private lateinit var postAuthor: String
-    private lateinit var postUpdated: String
     private lateinit var currentFeed: String
     private val commentsList = ArrayList<Comment>()
     private lateinit var modhash: String
@@ -67,7 +66,7 @@ class CommentsActivity : AppCompatActivity() {
 
          setupToolbar()
          getSessionParams()
-         setupImageLoader()
+         setupImageLoader(this)
          initPost()
          prepareCurrentFeed()
          makeGetFeedRequest()
@@ -100,7 +99,7 @@ class CommentsActivity : AppCompatActivity() {
     }
 
     private fun makeGetFeedRequest() {
-        val call = initRetrofit().getFeed(currentFeed)
+        val call = initXmlRetrofit().getFeed(currentFeed)
         call.enqueue(object : Callback<Feed> {
             override fun onResponse(call: Call<Feed>?, response: Response<Feed>?) {
                 val entries = response?.body()?.entrys
@@ -149,6 +148,10 @@ class CommentsActivity : AppCompatActivity() {
     }
 
     private fun initPost() {
+        lateinit var postThumbnailUrl: String
+        lateinit var postTitle: String
+        lateinit var postAuthor: String
+        lateinit var postUpdated: String
         intent.let {
             postUrl = it.getStringExtra(getString(R.string.post_url))
             postThumbnailUrl = it.getStringExtra(getString(R.string.post_thumbnail))
@@ -161,7 +164,7 @@ class CommentsActivity : AppCompatActivity() {
         postTitleTextView?.text = postTitle
         postAuthorTextView?.text = postAuthor
         postUpdatedTextView?.text = postUpdated
-        showImage(postThumbnailUrl, postThumbnailImageView, postLoadingProgressBar)
+        showImage(this ,postThumbnailUrl, postThumbnailImageView, postLoadingProgressBar)
     }
 
     private fun postReply() {
@@ -180,7 +183,7 @@ class CommentsActivity : AppCompatActivity() {
 
     private fun prepareCurrentFeed() {
         try {
-            val splittedUrl = postUrl.split(BASE_URL)
+            val splittedUrl = postUrl.split(BASE_URL + "r/")
             currentFeed = splittedUrl[1]
         } catch (e: ArrayIndexOutOfBoundsException) {
             Log.e(TAG, "initPost: ArrayIndexOutOfBoundsException: ${e.message}")
@@ -212,8 +215,7 @@ class CommentsActivity : AppCompatActivity() {
             headerMap["X-Modhash"] = modhash
             headerMap["cookie"] = "reddit_session=$cookie"
 
-            Log.e(TAG, "POST IDDDDDDDDDDDDDDDDDDDD: $postId + HEADER MAP: ${headerMap.toString()} + COMMENT : $comment")
-            val call = initRetrofitForComment().submitComment(headerMap, "comment", postId, comment)
+            val call = initGsonRetrofit().submitComment(headerMap, "comment", postId, comment)
             call.enqueue(object : Callback<CommentChecker> {
                 override fun onResponse(call: Call<CommentChecker>?, response: Response<CommentChecker>?) {
                     Log.e(TAG, "onResponse: SERVER RESPONSE: ${response.toString()}")
@@ -241,66 +243,6 @@ class CommentsActivity : AppCompatActivity() {
         Toast.makeText(this, "Error occured! Dud you sign in?", Toast.LENGTH_SHORT).show()
     }
 
-    private fun showImage(imageUrl: String, imageView: ImageView, progressBar: ProgressBar) {
-        val imageLoader = ImageLoader.getInstance()
-        val defaultImage = this.resources.getIdentifier("@drawable/reddit_alien", null, this.packageName)
-        val options = DisplayImageOptions.Builder().cacheInMemory(true)
-                .cacheOnDisk(true).resetViewBeforeLoading(true)
-                .showImageForEmptyUri(defaultImage)
-                .showImageOnFail(defaultImage)
-                .showImageOnLoading(defaultImage).build()
-
-        imageLoader.displayImage(imageUrl, imageView, options, object : ImageLoadingListener {
-            override fun onLoadingComplete(imageUri: String?, view: View?, loadedImage: Bitmap?) {
-                progressBar?.visibility = View.GONE
-            }
-
-            override fun onLoadingStarted(imageUri: String?, view: View?) {
-                progressBar?.visibility = View.VISIBLE
-            }
-
-            override fun onLoadingCancelled(imageUri: String?, view: View?) {
-                progressBar?.visibility = View.GONE
-            }
-
-            override fun onLoadingFailed(imageUri: String?, view: View?, failReason: FailReason?) {
-                progressBar?.visibility = View.GONE
-            }
-        })
-    }
-
-    private fun setupImageLoader() {
-        val defaultOptions = DisplayImageOptions.Builder()
-                .cacheOnDisk(true).cacheInMemory(true)
-                .imageScaleType(ImageScaleType.EXACTLY)
-                .displayer(FadeInBitmapDisplayer(300)).build()
-
-        val config = ImageLoaderConfiguration.Builder(this)
-                .defaultDisplayImageOptions(defaultOptions)
-                .memoryCache(WeakMemoryCache())
-                .discCacheSize(100 * 1024 * 1024).build()
-
-        ImageLoader.getInstance().init(config)
-    }
-
-    private fun initRetrofit() : FeedAPI {
-        val retrofit = Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(SimpleXmlConverterFactory.create())
-                .build()
-
-        return retrofit.create(FeedAPI::class.java)
-    }
-
-    private fun initRetrofitForComment() : FeedAPI {
-        val retrofit = Retrofit.Builder()
-                .baseUrl(COMMENT_URL)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-
-        return retrofit.create(FeedAPI::class.java)
-    }
-
     private fun onCommentClicked(comment: Comment) {
         getUserComment()
     }
@@ -312,6 +254,5 @@ class CommentsActivity : AppCompatActivity() {
              modhash = it.getString(getString(R.string.SessionModhash), "")
              cookie = it.getString(getString(R.string.SessionCookie), "")
          }
-         Log.e(TAG, "AAAAAAAAAAA: $username $modhash $cookie")
      }
 }
